Rename namespace from LArSoft to larsoft. Migrate package.py files.
