--- repo.yaml
@@ -1,2 +1,0 @@
-repo:
-  namespace: 'LArSoft'
