+++ spack-repo-index.yaml
@@ -1,0 +1,3 @@
+repo_index:
+  paths:
+  - spack_repo/larsoft
