+++ spack_repo/larsoft/packages/larcore/package.py
@@ -1,0 +1,64 @@
+# Copyright 2013-2019 Lawrence Livermore National Security, LLC and other
+# Spack Project Developers. See the top-level COPYRIGHT file for details.
+#
+# SPDX-License-Identifier: (Apache-2.0 OR MIT)
+
+from spack import *
+from spack.package import *
+from spack.util.prefix import Prefix
+from spack_repo.builtin.build_systems.cmake import CMakePackage
+from spack_repo.fnal_art.packages.fnal_github_package.package import FnalGithubPackage, cxxstd_variant
+
+
+class Larcore(CMakePackage, FnalGithubPackage):
+    """Larcore"""
+
+    repo = "LArSoft/larcore"
+    version_patterns = ["v09_00_00", "09.11.00"]
+
+    version("10.00.02", sha256="f9a614d5882ff40b4f6de172054e225e17f0460b50c74e673a70d008b6083c05")
+    version("10.00.00", sha256="be9c65b7a51d251ad167a1fcf0b51a11c0edcc992978a31c5fdedbcc2474c039")
+    version("09.11.01", sha256="60c85fab1f622274d6750c6578d372ad5c6d0fd28a857e79c82bcfcd828c6b09")
+    version("develop", branch="develop", get_full_repo=True)
+
+    cxxstd_variant("17", "20", default="17")
+
+    depends_on("c", type="build")
+    depends_on("cxx", type="build")
+
+    depends_on("cetmodules", type="build")
+
+    depends_on("art")
+    depends_on("art-root-io", "@:09.11.01")
+    depends_on("boost+test")
+    depends_on("cetlib")
+    depends_on("cetlib-except")
+    depends_on("fhicl-cpp")
+    depends_on("larcorealg")
+    depends_on("larcoreobj")
+    depends_on("messagefacility")
+
+    @cmake_preset
+    def cmake_args(self):
+        return [self.define_from_variant("CMAKE_CXX_STANDARD", "cxxstd")]
+
+    @sanitize_paths
+    def setup_build_environment(self, env):
+        prefix = Prefix(self.build_directory)
+        env.prepend_path("PATH", prefix.bin)  # Binaries.
+        env.prepend_path("CET_PLUGIN_PATH", prefix.lib)
+        env.prepend_path("FHICL_FILE_PATH", prefix.job)
+        env.prepend_path("FW_SEARCH_PATH", prefix.gdml)
+
+    @sanitize_paths
+    def setup_run_environment(self, env):
+        env.prepend_path("CET_PLUGIN_PATH", self.prefix.lib)
+        env.prepend_path("ROOT_INCLUDE_PATH", self.prefix.include)
+        env.prepend_path("FHICL_FILE_PATH", self.prefix.job)
+        env.prepend_path("FW_SEARCH_PATH", self.prefix.gdml)
+
+    def flag_handler(self, name, flags):
+        if name == "cxxflags" and self.spec.compiler.name == "gcc":
+            flags.append("-Wno-error=deprecated-declarations")
+            flags.append("-Wno-error=class-memaccess")
+        return (flags, None, None)
