+++ spack_repo/larsoft/packages/larfinder/findjsoncpp.patch
