+++ spack_repo/larsoft/packages/larreco/09.25.00.patch
