+++ spack_repo/larsoft/repo.yaml
@@ -1,0 +1,3 @@
+repo:
+  namespace: larsoft
+  api: v2.0
